requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
Pillow==10.2.0